export only libraryID == 1
I think libraryID = 1 is the "primary library". I've started using group
libraries, and if I export all, it upsets sorting (in Referey) by date
added (as the ones in the new library are much more recent, even if the original
paper was in my library long ago).

I guess more sophisticated schemes would be possible, like getting the
libraryID name, and putting that in a field of Mendeley. Not now, as I do not
need it.